# Conflicts:
#	README.md
#	pom.xml
#	threatconnect-app/apps-core/pom.xml
#	threatconnect-app/apps-util/pom.xml
#	threatconnect-app/playbooks-core/pom.xml
#	threatconnect-app/playbooks-test/pom.xml
#	threatconnect-app/pom.xml
#	threatconnect-sdk-addons/app-package-plugin/README.md
#	threatconnect-sdk-addons/app-package-plugin/pom.xml
#	threatconnect-sdk-addons/opendxl/pom.xml
#	threatconnect-sdk-addons/pom.xml
#	threatconnect-sdk-addons/stix-reader/pom.xml
#	threatconnect-sdk-addons/stix-writer/pom.xml
#	threatconnect-sdk-addons/threatconnect-parser/README.md
#	threatconnect-sdk-addons/threatconnect-parser/pom.xml
#	threatconnect-sdk/pom.xml
#	threatconnect-sdk/threatconnect-model/pom.xml
#	threatconnect-sdk/threatconnect-sdk-core/README.md
#	threatconnect-sdk/threatconnect-sdk-core/pom.xml

+++ threatconnect-sdk-addons/threatconnect-parser/src/main/java/com/threatconnect/sdk/parser/service/bulk/LegacyBulkIndicatorConverter.java
@@ -153,7 +153,7 @@ public class LegacyBulkIndicatorConverter
 	}
 	
 	public JsonObject convertToJson(final Collection<? extends Indicator> indicators,
-		final Map<Indicator, Set<Integer>> associatedIndicatorGroupsIDs)
+		final Map<Indicator, Set<Long>> associatedIndicatorGroupsIDs)
 	{
 		// create the root indicator json object
 		JsonObject indicatorJsonObject = new JsonObject();
@@ -165,7 +165,7 @@ public class LegacyBulkIndicatorConverter
 		// for each indicator in the collection
 		for (Indicator indicator : indicators)
 		{
-			Set<Integer> ids = associatedIndicatorGroupsIDs.get(indicator);
+			Set<Long> ids = associatedIndicatorGroupsIDs.get(indicator);
 			
 			try
 			{
@@ -181,7 +181,7 @@ public class LegacyBulkIndicatorConverter
 		return indicatorJsonObject;
 	}
 	
-	public JsonObject convertToJson(final Indicator indicator, final Set<Integer> associatedGroupIDs)
+	public JsonObject convertToJson(final Indicator indicator, final Set<Long> associatedGroupIDs)
 		throws InvalidIndicatorException
 	{
 		// holds the json object that will represent this indicator
@@ -201,7 +201,7 @@ public class LegacyBulkIndicatorConverter
 			JsonArray associatedGroups = new JsonArray();
 			
 			// for each of the ids
-			for (Integer id : associatedGroupIDs)
+			for (Long id : associatedGroupIDs)
 			{
 				associatedGroups.add(new JsonPrimitive(id));
 			}
