+++ threatconnect-sdk/threatconnect-sdk-core/src/main/java/com/threatconnect/sdk/server/entity/Owner.java
@@ -31,7 +31,7 @@ import javax.xml.bind.annotation.XmlSeeAlso;
 public abstract class Owner
 {
     @XmlElement(name = "Id", required = true)
-    private Integer id;
+    private Long id;
     @XmlElement(name = "Name", required = true)
     private String name;
     
@@ -42,19 +42,19 @@ public abstract class Owner
     {
     }
 
-    public Owner(Integer id, String name, String type)
+    public Owner(Long id, String name, String type)
     {
         this.id = id;
         this.name = name;
         this.type = type;
     }
 
-    public Integer getId()
+    public Long getId()
     {
         return id;
     }
 
-    public void setId(Integer id)
+    public void setId(Long id)
     {
         this.id = id;
     }
